notification is working
