+++ client/src/components/Layout/Frame/Frame.js
@@ -1,13 +1,17 @@
 import React from 'react';
+import Notification from '../../Notification/Notification';
 import { Container, Row } from 'react-bootstrap';
 
 const frame = props => {
     return (
-        <Container>
-            <Row>
-                {props.children}
-            </Row>
-        </Container>
+        <div>
+            <Notification />
+            <Container>
+                <Row>
+                    {props.children}
+                </Row>
+            </Container>
+        </div>
     )
 };
