--- client/src/components/Layout/Notification/Notification.js
@@ -1,13 +1,0 @@
-import React, { Fragment } from 'react';
-
-
-const notification = props => {
-
-    return (
-        <Fragment>
-
-        </Fragment>
-    );
-}
-
-export default notification;
