+++ client/src/components/Notification/Notification.css
@@ -1,0 +1,6 @@
+.notification {
+    float: right!important;
+    position: relative;
+    right: 10em;
+    top: 2em;
+}
