+++ client/src/controllers/DashboardController/DashboardController.js
@@ -4,10 +4,6 @@ import Frame from '../../components/Layout/Frame/Frame';
 import Card from '../../components/Layout/Cards/Cards';
 import Resource from '../Resource/Resource';
 
-import io from 'socket.io-client';
-
-import Notification from '../../components/Notification/Notification';
-
 class DashboardController extends Component {
     constructor(props) {
         super(props)
@@ -17,8 +13,6 @@ class DashboardController extends Component {
     state = {
         filter: 'all',
         search: '',
-        notifications: null,
-        socket_uri: 'http://localhost:3001'
     };
 
     handler(e, param) {
@@ -32,26 +26,6 @@ class DashboardController extends Component {
         })
     };
 
-    componentDidMount() {
-        let socket = io(this.state.socket_uri)
-
-        socket.on('notifications', data => {
-            this.setState({
-                notifications: data
-            })
-        })
-    }
-
-    componentDidUpdate() {
-        let socket = io(this.state.socket_uri)
-
-        socket.on('notifications', data => {
-            this.setState({
-                notifications: data
-            })
-        })
-    }
-
     render() {
         return (
             <Fragment>
@@ -60,7 +34,6 @@ class DashboardController extends Component {
                     handleClick={e => this.handler(e, 'click')}
                     show={true}
                 />
-                <Notification notification={this.state.notifications} />
                 <Frame>
                     <Resource
                         path={`/api/v1/documents?filter=${this.state.filter}`}
