+++ utils.js
@@ -23,6 +23,10 @@ module.exports = {
             month = `0${month}`
         }
 
+        if (day < 10) {
+            day = `0${day}`
+        }
+
         return `${year}-${month}-${day}`;
     },
 }
